converting beaconing model to new refactored class
